requirements.txt
# ─── SEMANTIC KERNEL ──────────────────────────────────────────
semantic-kernel>=1.5.0,<2.0.0

# ─── AZURE OPENAI ─────────────────────────────────────────────
openai>=1.35.0
azure-identity>=1.17.0

# ─── AZURE AI SEARCH (RAG) ────────────────────────────────────
azure-search-documents>=11.6.0

# ─── AZURE CONTENT SAFETY ─────────────────────────────────────
azure-ai-contentsafety>=1.0.0

# ─── REDIS CACHE FOR AZURE ────────────────────────────────────
redis[hiredis]>=5.0.0

# ─── API REST (FASTAPI) ───────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ─── SEGURIDAD JWT ────────────────────────────────────────────
PyJWT[crypto]>=2.8.0

# ─── HTTP CLIENT ──────────────────────────────────────────────
httpx>=0.27.0

# ─── UTILIDADES ───────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── TESTING ──────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
